Move Domain classes into subdirectory
