+++ Hackney.Shared.Asset/Domain/AssetAddress.cs
@@ -1,4 +1,4 @@
-namespace Hackney.Shared.Asset
+namespace Hackney.Shared.Asset.Domain
 {
     public class AssetAddress
     {
