+++ Hackney.Shared.Asset/Domain/AssetLocation.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Hackney.Shared.Asset
+namespace Hackney.Shared.Asset.Domain
 {
     public class AssetLocation
     {
